requirements.txt
Django==3.0.7
django-localflavor==3.0.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
requests==2.24.0
psycopg2-binary==2.8.5
python-dotenv==0.13.0
django-configurations==2.2
